+++ src/components/summary/CalorieSummary.tsx
@@ -1,0 +1,6 @@
+export const CalorieSummary = (props: { calories: number; }) => (
+  <div className="d-flex flex-column justify-content-center border rounded bg-info text-white text-center">
+    <div style={{ fontSize: '40px' }}>{props.calories}</div>
+    <div>calories</div>
+  </div>
+);
